Formatter: remove mistaken assert in dump_stream()
dump_stream was complaining if there was a pending string (basically,
an item still open), but this is standard usage for the dump_*() operations;
you start one, and then it's closed when the next one comes along.
Remove the assert and let print_spaces() close the existing element
and indent-if-pretty, as usual.