ceph_argparse.py: validate(): allow args to be lists
Multivalued arguments are passed in as lists.  Allow this, although
handle it somewhat cheesily by just extracting the first element
when encountered and pushing the rest back onto the front of the
arguments as non-list args.  This lets the looping/counting/validation
continue as normal.

Also, some comment cleanup and a nested-if removal because validate()
is looking pretty ugly.

Signed-off-by: Dan Mick <dan.mick@inktank.com>