hack/wip: experiment with turning off FOR_MAKE_CHECK
Why:
https://github.com/ceph/ceph/pull/64612/commits added a
python check-only dependency on python3-onelogin-saml2:"
https://github.com/ceph/ceph/pull/64612/files#diff-2a13f1344504379e877324d3a0375adcbcb4cb27d7e575ad8f4618a52d6a00e0

that package doesn't exist for focal.

that hasn't been a problem until now, because other build methods don't
use 'build-profile' pkg.ceph.check.  build-with-container.py does,
because it's trying to use the same container to build, potentially,
make check as it does for the actual package build.  because of that ^
PR, make check will no longer run on focal.

Here's a hack to turn off FOR_MAKE_CHECK to see how far we get.
(It allows the build container to be built, but the build fails in
C++ that I haven't root-caused yet).

Signed-off-by: Dan Mick <dan.mick@redhat.com>